Full Project Commit
Unfortunately, I did not realize I needed to commit to a GitHub
repository regularly. I did not see any guidance on the project
document. If it was mentioned in class, I may have missed it or been
absent that day. As such, I figured commiting my code to a public
repository would have been an honor code violation anyways. Anyhow,
after receiving guidance on Piazza indicating that you wanted a GitHub
link instead, I have decided to make a repository. In terms of
plagarism, all licenses and headers have been left in on copied
code/libraries. Anything that does not have a license on it is my own
work with the occasional google search 1-2 line snippet.